py/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.16.1
fsspec==2025.2.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
regex==2024.11.6
requests==2.32.3
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
